+++ sdk/servicebus/Azure.Messaging.ServiceBus/src/Client/ServiceBusClient.cs
@@ -385,63 +385,6 @@ namespace Azure.Messaging.ServiceBus
         }
 
         /// <summary>
-        /// Creates a <see cref="ServiceBusReceiver"/> instance that can be used for receiving from the
-        /// dead letter queue for the specified queue. It uses <see cref="ReceiveMode"/> to specify
-        /// how messages are received. Defaults to PeekLock mode. The <see cref="ReceiveMode"/> is set in <see cref="ServiceBusReceiverOptions"/>.
-        /// </summary>
-        ///
-        /// <param name="queueName">The queue to create a <see cref="ServiceBusReceiver"/> for.</param>
-        /// <param name="options">The set of <see cref="ServiceBusReceiverOptions"/> to use for configuring the
-        /// <see cref="ServiceBusReceiver"/>.</param>
-        ///
-        /// <returns>A <see cref="ServiceBusReceiver"/> scoped to the dead letter queue of the specified
-        /// queue.</returns>
-        public virtual ServiceBusReceiver CreateDeadLetterReceiver(
-            string queueName,
-            ServiceBusReceiverOptions options = default)
-        {
-            ValidateEntityName(queueName);
-
-            return new ServiceBusReceiver(
-                connection: Connection,
-                entityPath: EntityNameFormatter.FormatDeadLetterPath(queueName),
-                isSessionEntity: false,
-                plugins: Plugins,
-                options: options);
-        }
-
-        /// <summary>
-        /// Creates a <see cref="ServiceBusReceiver"/> instance that can be used for receiving from the
-        /// dead letter queue for the specified subscription. It uses <see cref="ReceiveMode"/> to specify
-        /// how messages are received. Defaults to PeekLock mode. The <see cref="ReceiveMode"/> is set in <see cref="ServiceBusReceiverOptions"/>.
-        /// </summary>
-        ///
-        /// <param name="topicName">The topic to create a <see cref="ServiceBusReceiver"/> for.</param>
-        /// <param name="subscriptionName">The subscription to create a <see cref="ServiceBusReceiver"/> for.</param>
-        /// <param name="options">The set of <see cref="ServiceBusReceiverOptions"/> to use for configuring the
-        /// <see cref="ServiceBusReceiver"/>.</param>
-        ///
-        /// <returns>A <see cref="ServiceBusReceiver"/> scoped to the dead letter queue of the specified
-        /// queue.</returns>
-        public virtual ServiceBusReceiver CreateDeadLetterReceiver(
-            string topicName,
-            string subscriptionName,
-            ServiceBusReceiverOptions options = default)
-        {
-            ValidateEntityName(topicName);
-
-            return new ServiceBusReceiver(
-                connection: Connection,
-                entityPath: EntityNameFormatter.FormatDeadLetterPath(
-                    EntityNameFormatter.FormatSubscriptionPath(
-                        topicName,
-                        subscriptionName)),
-                isSessionEntity: false,
-                plugins: Plugins,
-                options: options);
-        }
-
-        /// <summary>
         /// Creates a <see cref="ServiceBusProcessor"/> instance that can be used to process messages using
         /// event handlers that are set on the processor. It uses <see cref="ReceiveMode"/> to specify
         /// how messages are received. Defaults to PeekLock mode. If you want to change the <see cref="ReceiveMode"/>,
@@ -631,8 +574,7 @@ namespace Azure.Messaging.ServiceBus
         /// Validates that the specified entity name matches the entity path in the Connection,
         /// if an entity path is specified in the connection.
         /// </summary>
-        /// <param name="entityName">Entity name to validate</param>
-        ///
+        /// <param name="entityName">Entity name to validate.</param>
         /// <param name="sendViaEntityName">The send via entity name to validate</param>
         private void ValidateSendViaEntityName(string entityName, string sendViaEntityName)
         {