+++ sdk/keyvault/Azure.Security.KeyVault.Keys/CHANGELOG.md
@@ -1,10 +1,15 @@
 # Release History
 
-## 4.2.0-preview.1 (Unreleased)
+## 4.2.0-beta.1 (Unreleased)
 
 ### Added
 
 - Added `KeyVaultKeyIdentifier` to parse key URIs.
+- Added `LocalCryptographyClient` to do cryptography operations locally using a `JsonWebKey`.
+
+### Changed
+
+- Clarified documentation for `KeyClientBuilderExtensions` methods.
 
 ## 4.1.0 (2020-08-11)